+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Features:
 
+- Validators functions available from Nelson and C++ API (part 2):
+
+  - `mustBeFile`
+
 - `test_run` manages `SEQUENTIAL TEST REQUIRED` and `NATIVE_ARCHITECTURE TEST REQUIRED` tags.
+
 - benchs are executed sequentialy (better bench results).
 
 ## Bug Fixes: